requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
diff-match-patch==20241021
dj-database-url==3.0.1
Django==4.2.11
django-cors-headers==4.7.0
django-filter==23.5
django-import-export==4.3.14
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
kombu==5.6.2
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.1.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
